+++ db/updates/update_11.0-11.1.sql
@@ -1,3 +1,0 @@
-ALTER TABLE `batchIDHistory` CHANGE `id` `id` VARCHAR(255) NOT NULL; 
-ALTER TABLE `settings` ADD `currency_code` VARCHAR(255) NOT NULL DEFAULT 'GBP' AFTER `currency`; 
-ALTER TABLE `settings` CHANGE `currency` `currency` VARCHAR(255) NOT NULL DEFAULT '£';